Fix miscellaneous compilation warnings

+++ demos/sphere/demo_sphere_irreg_waves_eta_import.cpp
@@ -108,7 +108,7 @@ int main(int argc, char* argv[]) {
     std::cout << "Defining irregular wave input parameters..." << std::endl;
     IrregularWaveParams params;
     std::cout << "bodies.size() = " << bodies.size() << std::endl;
-    params.num_bodies_          = bodies.size();
+    params.num_bodies_          = (unsigned int)bodies.size();
     params.simulation_dt_       = timestep;
     params.simulation_duration_ = simulationDuration;
     params.ramp_duration_       = 0.0;
